move descriptor based stuff to test folder
