--- src/vslam/include/vslam/descriptor_matching.h
@@ -1,24 +1,0 @@
-// Copyright 2022 Philipp.Duernay
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License
-// along with this program.  If not, see <https://www.gnu.org/licenses/>.
-
-#ifndef VSLAM_DESCRIPTOR_MATCHING_PACKAGE_H__
-#define VSLAM_DESCRIPTOR_MATCHING_PACKAGE_H__
-#include "descriptor_matching/FeatureTracking.h"
-#include "descriptor_matching/FeatureTrackingOcv.h"
-#include "descriptor_matching/Matcher.h"
-#include "descriptor_matching/bundle_adjustment/BundleAdjustment.h"
-#include "descriptor_matching/overlays.h"
-
-#endif  // VSLAM_DESCRIPTOR_MATCHING_PACKAGE_H__
