+++ src/vslam/test/descriptor_matching/test_matching.cpp
@@ -19,8 +19,12 @@
 
 #include <gtest/gtest.h>
 
+#include "BundleAdjustment.h"
+#include "FeatureTracking.h"
+#include "FeatureTrackingOcv.h"
+#include "Matcher.h"
+#include "overlays.h"
 #include "vslam/core.h"
-#include "vslam/descriptor_matching.h"
 #include "vslam/evaluation.h"
 #include "vslam/features.h"
 #include "vslam/utils.h"
